backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
shippo>=3.1.0  # Updated to allow latest compatible version
python-dotenv==1.0.1
pydantic==2.6.0